Move ConfigurableColumnsHeader to a separate file
