+++ picard/ui/itemviews/basetreeview.py
@@ -88,92 +88,13 @@ from picard.util import (
 )
 
 from picard.ui.collectionmenu import CollectionMenu
-from picard.ui.columns import ImageColumn
 from picard.ui.enums import MainAction
 from picard.ui.ratingwidget import RatingWidget
 from picard.ui.scriptsmenu import ScriptsMenu
 from picard.ui.util import menu_builder
-from picard.ui.widgets.tristatesortheaderview import TristateSortHeaderView
-
-
-class ConfigurableColumnsHeader(TristateSortHeaderView):
-
-    def __init__(self, columns, parent=None):
-        super().__init__(QtCore.Qt.Orientation.Horizontal, parent)
-        self._columns = columns
-        self._always_visible_columns = set(self._columns.always_visible_columns())
-        self._visible_columns = set(self._always_visible_columns)
-
-        self.sortIndicatorChanged.connect(self.on_sort_indicator_changed)
-
-        # enable sorting, but don't actually use it by default
-        # XXX it would be nice to be able to go to the 'no sort' mode, but the
-        #     internal model that QTreeWidget uses doesn't support it
-        self.setSortIndicator(-1, QtCore.Qt.SortOrder.AscendingOrder)
-
-    def show_column(self, column, show):
-        if column in self._always_visible_columns:
-            # Always visible
-            # Still execute following to ensure it is shown
-            show = True
-        self.parent().setColumnHidden(column, not show)
-        if show:
-            self._visible_columns.add(column)
-        else:
-            self._visible_columns.discard(column)
-
-    def contextMenuEvent(self, event):
-        menu = QtWidgets.QMenu(self)
-        parent = self.parent()
-
-        for i, column in enumerate(self._columns):
-            if i in self._always_visible_columns:
-                continue
-            action = QtGui.QAction(_(column.title), parent)
-            action.setCheckable(True)
-            action.setChecked(i in self._visible_columns)
-            action.setEnabled(not self.is_locked)
-            action.triggered.connect(partial(self.show_column, i))
-            menu.addAction(action)
-
-        menu.addSeparator()
-        restore_action = QtGui.QAction(_("Restore default columns"), parent)
-        restore_action.setEnabled(not self.is_locked)
-        restore_action.triggered.connect(self.restore_defaults)
-        menu.addAction(restore_action)
-
-        lock_action = QtGui.QAction(_("Lock columns"), parent)
-        lock_action.setCheckable(True)
-        lock_action.setChecked(self.is_locked)
-        lock_action.toggled.connect(self.lock)
-        menu.addAction(lock_action)
-
-        menu.exec(event.globalPos())
-        event.accept()
-
-    def restore_defaults(self):
-        self.parent().restore_default_columns()
-
-    def paintSection(self, painter, rect, index):
-        column = self._columns[index]
-        if isinstance(column, ImageColumn):
-            painter.save()
-            super().paintSection(painter, rect, index)
-            painter.restore()
-            column.paint(painter, rect)
-        else:
-            super().paintSection(painter, rect, index)
-
-    def on_sort_indicator_changed(self, index, order):
-        if not self._columns[index].sortable:
-            self.setSortIndicator(-1, QtCore.Qt.SortOrder.AscendingOrder)
-
-    def lock(self, is_locked):
-        super().lock(is_locked)
-
-    def __str__(self):
-        name = getattr(self.parent(), 'NAME', str(self.parent().__class__.__name__))
-        return f"{name}'s header"
+from picard.ui.widgets.configurablecolumnsheader import (
+    ConfigurableColumnsHeader,
+)
 
 
 def _alternative_versions(album):
